Renames logger directory to loggers
